remove mouse param for Context::new
You don't have access to the mouse where you want to create a context.
As a result, even in the example it was set to 0,0.
This makes no sense, so now the start at 0,0 is done for you.

I may add a new method that still takes a mouse location in the future.